requirements.txt
altair==5.2.0
appdirs==1.4.4
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
biocommons.seqrepo==0.6.5
biopython==1.83
biothings-client==0.3.1
bioutils==0.5.8.post1
blinker==1.7.0
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
configparser==6.0.1
cssselect==1.2.0
decorator==5.1.1
diskcache==5.6.3
executing==2.0.1
fake-useragent==1.4.0
filelock==3.13.1
genomepy==0.16.1
gitdb==4.0.11
GitPython==3.1.42
hgvs==1.5.1
humanfriendly==10.0
idna==3.6
importlib-metadata==7.0.1
importlib_resources==6.1.2
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
loguru==0.7.2
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
mygene==3.2.2
mysql-connector-python==8.3.0
norns==0.1.6
nose==1.3.7
numpy==1.24.4
packaging==23.2
pandas==2.0.3
parse==1.20.1
Parsley==1.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pkgutil_resolve_name==1.3.10
primer3-py==2.0.3
prompt-toolkit==3.0.43
protobuf==4.25.3
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pydeck==0.8.1b0
pyee==11.1.0
pyfaidx==0.8.1.1
Pygments==2.17.2
pyppeteer==2.0.0
pyquery==2.0.0
pysam==0.22.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-html==0.10.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
streamlit==1.36.0
tabulate==0.9.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.18
validators==0.22.0
w3lib==2.1.2
watchdog==4.0.0
wcwidth==0.2.13
websockets==10.4
yoyo-migrations==8.2.0
zipp==3.17.0
